move method for creating item node to appropriate module
