+++ src/itemNodes.js
@@ -1,9 +1,10 @@
 const { parse, stringify } = require(`flatted/cjs`);
+const _ = require('lodash');
+const changeCase = require('change-case');
 
+const normalize = require('./normalize');
 const richTextElementDecorator =
   require('./decorators/richTextElementDecorator');
-const normalize = require(`./normalize`);
-
 
 /**
  * Creates an array of content item nodes in default culture
@@ -30,8 +31,11 @@ const getFromDefaultLanguage = async (
   const itemsFlatted = parse(stringify(contentItemsResponse.items));
   const contentItemNodes = itemsFlatted.map((contentItem) => {
     try {
-      return normalize
-        .createContentItemNode(createNodeId, contentItem, contentTypeNodes);
+      return createContentItemNode(
+        createNodeId,
+        contentItem,
+        contentTypeNodes
+      );
     } catch (error) {
       console.error(error);
     }
@@ -66,7 +70,7 @@ const getFromNonDefaultLanguage = async (
 
     const languageItemsFlatted = parse(stringify(languageResponse.items));
     const contentItemsNodes = languageItemsFlatted.map((languageItem) =>
-      normalize.createContentItemNode(
+      createContentItemNode(
         createNodeId,
         languageItem,
         contentTypeNodes
@@ -76,6 +80,56 @@ const getFromNonDefaultLanguage = async (
   };
   return nonDefaultLanguageItemNodes;
 };
+
+/**
+ * Creates a Gatsby object out of a Kentico Cloud content item object.
+ * @param {function} createNodeId - Gatsby function to create a node ID.
+ * @param {object} contentItem - Kentico Cloud content item object.
+ * @param {array} contentTypeNodes - All Gatsby content type nodes.
+ * @return {object} Gatsby content item node.
+ * @throws {Error}
+ */
+const createContentItemNode =
+  (createNodeId, contentItem, contentTypeNodes) => {
+    if (typeof createNodeId !== `function`) {
+      throw new Error(`createNodeId is not a function.`);
+    } else if (!contentItem || !_.has(contentItem, `system.codename`)) {
+      throw new Error(`contentItem is not a valid content item object.`);
+    } else if (!contentTypeNodes
+      || !_.isArray(contentTypeNodes)
+      || (!_.isEmpty(contentTypeNodes)
+        && !_.has(contentTypeNodes, `[0].system.codename`))) {
+      throw new Error(`contentTypeNodes is not an array of valid objects.`);
+    }
+    const codenameParamCase =
+      changeCase.paramCase(contentItem.system.codename);
+
+    const languageParamCase =
+      changeCase.paramCase(contentItem.system.language);
+
+    const nodeId = createNodeId(
+      `kentico-cloud-item-${codenameParamCase}-${languageParamCase}`
+    );
+
+    const parentContentTypeNode = contentTypeNodes.find(
+      (contentType) => contentType.system.codename
+        === contentItem.system.type);
+
+    const itemWithElements = normalize.parseContentItemContents(contentItem);
+
+    const additionalData = {
+      otherLanguages___NODE: [],
+      contentType___NODE: parentContentTypeNode.id,
+    };
+
+    return normalize.createKcArtifactNode(
+      nodeId,
+      itemWithElements,
+      `item`,
+      contentItem.system.type,
+      additionalData
+    );
+  };
 
 module.exports = {
   getFromDefaultLanguage,
